Added Yearly Average to All Monitors
Yearly average added to each agent row and fixed a bug that prevented needed monitors to show up if the result object is empty. Code was set to (result != ""), but since it is an object, the empty string check failed. Changed to (result) to check for an empty object. Upped the Version Number